refactor: viewModelFactory 패키지 이동
